Add files via upload
AND operator:
In MongoDB, the AND operation is implicit when you specify multiple key-value pairs in a query object.
OR operator:
The OR operator in MongoDB is represented using the $or operator. It allows you to specify multiple conditions, any of which can match for a document to be included in the results.
Combining AND and OR (AND/OR):
You can combine AND and OR operators to create more complex queries.
Comparison Operators:
$eq: Matches values that are equal to a specified value.
$ne: Matches all values that are not equal to a specified value.
$gt: Matches values that are greater than a specified value.
$gte: Matches values that are greater than or equal to a specified value.
$lt: Matches values that are less than a specified value.
$lte: Matches values that are less than or equal to a specified value.
Logical Operators:
$and: Joins query clauses with a logical AND.
$or: Joins query clauses with a logical OR.
$not: Inverts the effect of a query expression and returns documents that do not match the query expression.
$nor: Joins query clauses with a logical NOR, which returns documents that fail all the specified conditions.
insertOne() - This inserts a single document with the specified fields into the collection.
insertMany() - insertMany() is used to insert multiple documents into the collection in a single operation.
deleteMany()- deletes all documents that match the specified condition.
deleteOne() - This deletes a single document that matches the specified condition.